refactor(racingcar): 코드 리뷰 의견 적용
