+++ src/main/java/racingcar/MoveCondition.java
@@ -5,15 +5,3 @@ public interface MoveCondition {
     boolean movable();
 }
 
-class NumberMoveCondition implements MoveCondition {
-    private final int condition;
-
-    public NumberMoveCondition(int condition) {
-        this.condition = condition;
-    }
-
-    @Override
-    public boolean movable() {
-        return this.condition >= 4;
-    }
-}
